cleanup + add editorconfig

+++ core/src/main/java/de/gerrygames/viarewind/protocol/protocol1_7_6_10to1_8/packets/PlayerPackets.java
@@ -95,7 +95,7 @@ public class PlayerPackets {
 				handler(packetWrapper -> {
 					Position position = packetWrapper.read(Type.POSITION);
 					packetWrapper.write(Type.INT, position.getX());
-					packetWrapper.write(Type.INT, (int) position.getY());
+					packetWrapper.write(Type.INT, position.getY());
 					packetWrapper.write(Type.INT, position.getZ());
 				});
 			}
@@ -162,7 +162,7 @@ public class PlayerPackets {
 						y += playerPosition.getPosY();
 					}
 					playerPosition.setReceivedPosY(y);
-					y += (double) 1.62F;
+					y += 1.62F;
 					packetWrapper.set(Type.DOUBLE, 1, y);
 					if ((flags & 0x04) == 0x04) {
 						double z = packetWrapper.get(Type.DOUBLE, 2);
